Move `load_ui()` helper to dedicated utils module
